Do not run destreamer as privileged user
chromium does not run as root/admin unless its sandbox is disabled.

To prevent security breaches, we will not support running as privileged user.

Everyone is free to make the necessary changes to remove this restriction, in which case destreamer will not be responsible for any issue or data loss.

+ Update README