Add option to change encoder and container
This change enables the user to select a specific
container and/or encoder for the output. As this change
forces ffmpeg to use copy by default, destreamer now uses
mkv as default output. This might be needed when MS Streams
doesn't offer a mp4 compatible stream. mkv offers much more
support.