Don't wait for email field if username isn't provided (#123)
Destreamer waits for the `<input type="email">` tag to appear. But in
some corporate environments, like mine, the videos are protected by
proxies, which in turn ar protected by SSO/SAML. Since the SSO login may
not have that input field's type set to 'email', the wait would
eventually timeout.

With this fix we wait for the "email" field only if --username option
was used. Otherwise we let the user take the actions manually to take
the browser to the video page.